pull next_race code from main.py
